https: rewrite certbot plugin.
Currently the certbot plugin used in the snap is a copy of the webroot
plugin from upstream certbot with the chown bits commented out. However,
it turns out that the webroot plugin as-written cannot be used as an
external plugin. It's far more complicated than we needed anyway, so
rewrite it to be exactly what we need.

Fix #159.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>