many: place all pidfiles and sockets in /tmp.
Until recently, /tmp within a snap was unique to the app in the snap.
This meant that any sockets or pidfiles placed into /tmp in app A was
not accessible from app B (as it saw a separate /tmp).

As of snap-confine v1.0.43, this is no longer the case. /tmp is now
unique snap-wide: /tmp for app A is the same /tmp as app B as long as
both apps are contained within the same snap.

This finally gives us a place accessible from confinement that's wiped
on each boot. It's the place to put pidfiles and sockets so hard reboots
don't cause future failures.

Fix #151.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>