apache: support changing ports
Currently, Apache listens on port 80. If HTTPS is enabled, it listens on
both 80 and 443. This is inflexible, especially if one wants to put the
snap behind a proxy of some kind.

Support changing HTTP and HTTPS ports via the `configure` hook.
Expose these configurations via the `ports.http` and `ports.https`
options, respectively.

Fix #134.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>