php: enable ldap support
Enable the LDAP module in PHP, allowing the LDAP app to be enabled. This
requires a slight hack to replace ARCH_TRIPLET in the flags. This is
similar to how Snapcraft v2.36 works, so once that's released we can
remove this small hack and start using SNAPCRAFT_ARCH_TRIPLET instead.

Fix #117

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>