occ: don't require PHP restart for maintenance mode
The `nextcloud.occ` command is often used to alter the Nextcloud config,
which is a .php file. As such, it's subject to the same opcache settings
as the rest of PHP, which doesn't even look at the tree to see if files
have changed (`opcache.validate_timestamps=0`). As a result, any change
made to `config.php` without restarting PHP goes unnoticed, which
includes changes like enabling maintenance mode.

When using `nextcloud.occ`, assume the config has changed, and
invalidate the opcache for the config so that it's reloaded.

Fix #486

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>