nextcloud: notify admins upon update
The main sell of the Nextcloud snap is that is updates automatically in
a transactional manner. A side effect of this is that admins could
completely miss the fact that an update occurred, and if the update did
something they didn't expect they don't immediately know what to blame.

Notify all admins when the snap updates, making sure they know how to
revert if something doesn't work properly.

Resolve #1220

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>