hooks: add pre-refresh hook to update apps
Nextcloud's update process doesn't do a good job of ensuring apps are
updated properly and re-enabled after the update. It seems to work best
if apps are already as up-to-date as they can be. If we know we're about
to be refreshed, take a moment to make sure all apps are up-to-date.

Resolve #1243

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>