mysql: ignore config files on host
Snaps are not properly confined across distributions. This has led to
bad behavior in the Nextcloud snap where, if installed on a Debian
system (which does not have apparmor in its kernel) that has MySQL
installed, the snap will find the host's MySQL configuration files. This
isn't an issue on Ubuntu or other systems where snap security is real.

Work around this by explicitly not loading any config files except the
one in the snap.

Resolve #913

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>