migrations: sanity check nextcloud version before upgrade
Nextcloud is strict about what makes for a valid upgrade. Minor version
jumps are okay (i.e. within the same major version). Single major
version jumps are also okay, but not more than one (e.g. 18 to 19 is
okay, but not 18 to 20). Downgrades are never okay.

Add a check to the migration runner (which is run by the post-refresh
hook) to ensure that the snap doesn't allow such a refresh to occur,
which should relieve a lot of potential stress and reverts.

Resolve #1393

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>