integration tests: wait for nextcloud fixer to finish
We have some tests that reinstall the snap, in which case Nextcloud
Fixer runs, which is something that causes maintenance mode to be
enabled. If the timing is Just Right (or Wrong, as the case may be),
maintenance mode will be enabled right as a test is trying to log in,
which causes the test to fail. Just wait for the Fixer to finish before
trying anything.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>