ci: publish snaps from pull requests for testing
Take the snap created/tested in a pull request, and upload it to a
channel following the naming pattern `latest/beta/pr-<PR ID>`. This
works even on distrusted forks due to the use of the `workflow_run`
trigger. Once the pull request's CI finishes successfully, this job is
triggered on the _master_ branch, which is trusted code. The job
downloads the snap artifact from the pull request, and handles the
upload and token access without ever running distrusted code. This keeps
the risk of token leakage to a minimum.

Note that this only releases the snap built during the pull request's CI
process, which is amd64.

Resolve #2521.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>