PR snap releaser: detect triggering pull request in fork-compatible way
2df9c560119a2544774f11101006d9d25ab67cb1 used the Workflow Run info in
the GitHub API to determine the pull request that triggered it. Sadly
that information doesn't appear to be filled out properly if the pull
request in question is coming from a fork. Use a more brute-force method
instead, searching through every open pull request for the commit that
triggered the workflow in question.

Resolve #2521

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>